+++ src/Separador.js
@@ -1,0 +1,13 @@
+function separarCodigo(cod1) {
+
+    let numString = String(cod1);
+    const array = []    
+
+    for(var i = 0;i<numString.length;i++) {
+        let valInsert = +numString.charAt(i);
+        array.push(valInsert);        
+    }
+    return array;
+}
+
+export default separarCodigo;
